+++ src/rpbp/analysis/profile_construction/summarize_rpbp_profile_construction.py
@@ -20,8 +20,8 @@ import pbiotools.misc.parallel as parallel
 import pbiotools.misc.shell_utils as shell_utils
 import pbiotools.misc.utils as utils
 
-import pbiotools.ribo.ribo_filenames as filenames
-import pbiotools.ribo.ribo_utils as ribo_utils
+import rpbp.ribo_utils.utils as ribo_utils
+import rpbp.ribo_utils.filenames as filenames
 
 from rpbp.defaults import default_num_cpus, metagene_options