+++ src/rpbp/translation_prediction/estimate_orf_bayes_factors.py
@@ -17,7 +17,7 @@ import pbiotools.misc.parallel as parallel
 import pbiotools.misc.slurm as slurm
 import pbiotools.misc.utils as utils
 
-import pbiotools.ribo.ribo_utils as ribo_utils
+import rpbp.ribo_utils.utils as ribo_utils
 
 from cmdstanpy import CmdStanModel
 from rpbp.defaults import default_num_groups, translation_options